requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
bcrypt==4.2.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.4
cryptography==44.0.0
dill==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==33.3.0
fastapi==0.115.4
greenlet==3.1.1
h11>= 0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==24.2
passlib==1.7.4
platformdirs==4.3.6
playwright==1.48.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pyee==12.0.0
PyJWT==2.10.1
pylint==3.3.1
pytest==8.3.3
pytest-cov==6.0.0
pytest-pylint==0.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tenacity==9.0.0
tomli==2.0.2
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
